requirements.txt
pytest-playwright
playwright
pyclean
pytest-xdist